requirements.txt
flask>=2.2.2
python-dotenv>=0.21.0
requests>=2.28.1
azure-keyvault-secrets>=4.7.0
azure-identity>=1.12.0
firebase_admin>=6.2.0
PyYAML>=6.0.1